Fixes for eapol_test in case of eap_sim, eap_aka, eap_aka_prime
Added a condition which checks for existence of OPc value and only then copies to keys.auc.opc. Additionally,
initialized the opc_p variable to NULL as it was always having some garbage values resulting in triggering the
condition added in this commit even in cases where OPc was not present.